requirements.txt
openai
python-dotenv
jinja2
PyYAML
beautifulsoup4
requests
